+++ inst/examples/parseYaml.r
@@ -1,9 +1,13 @@
 \dontrun{
 
+##------------------------------------------------------------------------------
+## From function //
+##------------------------------------------------------------------------------
+
 ## Get //
 yaml <- getYaml(
   from = function() {
-  "object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
+    "object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
   },
   ctx = YamlContext.ObjectReference.S3()
 )
@@ -18,7 +22,27 @@ yaml$parsed
 yaml$index
 yaml$src
 
+## Condition handling //
+yaml <- getYaml(
+  from = function() {
+    "Invalid YAML markup"
+  },
+  ctx = YamlContext.ObjectReference.S3()
+)
+yaml$original
+
+try(getYaml(
+  from = function() {
+    "Invalid YAML markup"
+  },
+  ctx = YamlContext.ObjectReference.S3(),
+  strict = TRUE
+))
+
+##------------------------------------------------------------------------------
 ## From expression //
+##------------------------------------------------------------------------------
+
 yaml <- getYaml(
   from = captureExpression({
     "object-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"